+++ fit-as-a-fiddle/app/filter/ResultsDisplayGrid.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import MedicationInfoCard from './MedicationInfoCard';
+import MedicationInfoCard from '../search/MedicationInfoCard';
 import { Container, Row, Col } from 'react-bootstrap';
 import Pagination from 'react-bootstrap/Pagination';
 
@@ -7,7 +7,7 @@ const chunkSize = 9; // Number of cards per page
 
 const ResultsDisplayGrid = ({ medications }) => {
   const [currentPage, setCurrentPage] = useState(0);
-  
+
   // Flatten the 2D array
   const flatMedications = medications.reduce((acc, val) => acc.concat(val), []);
 
@@ -19,10 +19,10 @@ const ResultsDisplayGrid = ({ medications }) => {
     }
     return result;
   };
-  
+
   // Creating chunks/pages of data
   const pages = chunkArray(flatMedications, chunkSize);
-  
+
   // The data for the current page
   const currentData = pages[currentPage] || [];
